fix(auction): tighten lock-test assertion and move proxy-bid index initializer to config package
